+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Support mentions containing internationalized domain names.
+- Add `gateways` property to portable actor objects.
 
 ### Changed